pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-template"
version = "1.0.0"
requires-python = ">=3.13"
dependencies = [
    "bcrypt",
    "fastapi[standard]",
    "psycopg2-binary",
    "pydantic-settings",
    "pyjwt",
    "sqlmodel",
]

[project.scripts]
api-template = "main:main"

[dependency-groups]
data = ["alembic"]
docs = ["mkdocs-material"]
lint = ["ruff==0.12.0"]
test = ["pytest-cov", "pytest-playwright", "testcontainers"]

[tool.hatch.build]
include = ["main.py", "src/**/*.py"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["test"]
addopts = ["--cov=main", "--cov=src", "--cov-report=html", "--cov-report=term"]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.format]
indent-style = "space"
quote-style = "double"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812", # missing-trailing-comma
    "D100",   # undocumented-public-module
    "D101",   # undocumented-public-class
    "D102",   # undocumented-public-method
    "D103",   # undocumented-public-function
    "D107",   # undocumented-public-init
    "D203",   # incorrect-blank-line-before-class
    "D213",   # multi-line-summary-second-line
    "INP001", # implicit-namespace-package
]

[tool.ruff.lint.isort]
known-third-party = ["alembic"]

[tool.ruff.lint.per-file-ignores]
"*test*.py" = [
    "PLR2004", # magic-value-comparison
    "S101",    # assert
]
